+++ Model/Operation/AttributeOptionsOperation.php
@@ -25,8 +25,10 @@ use Magento\Eav\Model\ResourceModel\Entity\Attribute\Option;
 use Magento\Eav\Model\ResourceModel\Entity\Attribute\Option\Collection;
 use Magento\Eav\Model\ResourceModel\Entity\Attribute\Option\CollectionFactory;
 use Magento\Framework\App\Config\ScopeConfigInterface;
+use Magento\Framework\Exception\InputException;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Framework\Phrase;
 use Magento\Store\Model\StoreManagerInterface;
 
 use function __;
@@ -77,6 +79,7 @@ class AttributeOptionsOperation implements AttributeOptionsInterface
     protected $attributeOptionCollectionFactory;
 
     //phpcs:ignoreFile Generic.Files.LineLength.Too.Long
+
     /**
      * @param \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig
      * @param \Magento\Eav\Api\AttributeRepositoryInterface $attributeRepository
@@ -149,6 +152,7 @@ class AttributeOptionsOperation implements AttributeOptionsInterface
         }
 
         $newOptionList = [];
+        $attributeOptionErrors = [];
 
         foreach ($newOptions as $option) {
             $newOptionList[$option->getAdminValue()] = $option;
@@ -160,15 +164,41 @@ class AttributeOptionsOperation implements AttributeOptionsInterface
         foreach ($currentOptions as $currentOption) {
             /** @var \Magento\Eav\Model\Entity\Attribute\Option $currentOption */
             if (array_key_exists($currentOption->getValue(), $newOptionList)) {
-                $updatedOption = $newOptionList[$currentOption->getValue()];
-                $this->updateOption($attribute, $currentOption, $updatedOption);
-                unset($newOptionList[$currentOption->getValue()]);
+                try {
+                    $updatedOption = $newOptionList[$currentOption->getValue()];
+                    $this->updateOption($attribute, $currentOption, $updatedOption);
+                    unset($newOptionList[$currentOption->getValue()]);
+                } catch (InputException $exception) {
+                    $attributeOptionErrors[] = __(
+                        'An error occured while importing value(' .
+                        $updatedOption->getAdminValue() .
+                        ') for attribute(' .
+                        $attribute->getAttributeCode() . '): ' .
+                        $exception->getMessage()
+                    );
+                }
             }
         }
 
         // Create remaining options
         foreach ($newOptionList as $newOption) {
-            $this->createOption($attribute, $newOption);
+            try {
+                $this->createOption($attribute, $newOption);
+            } catch (InputException $exception) {
+                $attributeOptionErrors[] = __(
+                    'An error occured while importing value(' .
+                    $newOption->getAdminValue() .
+                    ') for attribute(' .
+                    $attribute->getAttributeCode() . '): ' .
+                    $exception->getMessage()
+                );
+            }
+        }
+
+        if (count($attributeOptionErrors) >0) {
+            $errorMessage = __('There were errors while importing attributes option values: ');
+            $errorMessage .= implode(',', $attributeOptionErrors);
+            throw new InputException(new Phrase($errorMessage));
         }
     }
 
@@ -185,12 +215,13 @@ class AttributeOptionsOperation implements AttributeOptionsInterface
         AttributeInterface $attribute,
         AttributeOption $currentOption,
         OptionInterface $updatedOption
-    ): void {
+    ): void
+    {
         $labels = $this->getStoreLabels($updatedOption);
         $currentOption->setStoreLabels($labels);
 
         $this->attributeOptionManagement->update(
-            (int) $attribute->getEntityTypeId(),
+            (int)$attribute->getEntityTypeId(),
             $attribute->getAttributeCode(),
             $currentOption
         );